+++ AppboyKit/headers/AppboyKitLibrary/ABKCard.h
@@ -50,8 +50,17 @@ NS_ASSUME_NONNULL_BEGIN
  */
 @property (copy, nullable) NSString *urlString;
 
+/*!
+ * When the card's urlString is not nil, if the property is set to YES, the URL will be opened in a modal UIWebView
+ * inside the app. If this property is set to NO, the URL will be opened by the OS and web URLs will be opened in
+ * an external web browser app.
+ *
+ * This property defaults to NO.
+ */
+@property BOOL openUrlInWebView;
+
 /*
- * @param cardDictionary The dictionary for card deserialization.`
+ * @param cardDictionary The dictionary for card deserialization.
  *
  * Deserializes the dictionary to a card for use by wrappers such as Appboy's Unity SDK for iOS.
  * When the deserialization isn't successful, this method returns nil; otherwise, it returns the deserialized card.