+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageButton.h
@@ -38,6 +38,15 @@ NS_ASSUME_NONNULL_BEGIN
 @property (readonly, copy, nullable) NSURL *buttonClickedURI;
 
 /*!
+ * When the button's buttonClickActionType is ABKInAppMessageRedirectToURI, if the property is set to YES,
+ * the URI will be opened in a modal UIWebView inside the app. If this property is set to NO, the URI will be opened by
+ * the OS and web URIs will be opened in an external web browser app.
+ * 
+ * This property defaults to NO.
+ */
+@property BOOL buttonOpenUrlInWebView;
+
+/*!
  * This property defines the button's ID. Button's ID is used to track user's clicking action and used for corresponding
  * data analytics.
  */