+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageFullViewController.h
@@ -5,5 +5,6 @@
  */
 NS_ASSUME_NONNULL_BEGIN
 @interface ABKInAppMessageFullViewController : ABKInAppMessageImmersiveViewController
+
 @end
 NS_ASSUME_NONNULL_END