+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageHTMLFullViewController.h
@@ -2,5 +2,6 @@
 
 NS_ASSUME_NONNULL_BEGIN
 @interface ABKInAppMessageHTMLFullViewController : ABKInAppMessageHTMLViewController
+
 @end
 NS_ASSUME_NONNULL_END