+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageImmersiveViewController.h
@@ -8,7 +8,7 @@
 NS_ASSUME_NONNULL_BEGIN
 @interface ABKInAppMessageImmersiveViewController : ABKInAppMessageViewController
 
-@property (weak) IBOutlet ABKLabel *inAppMessageHeaderLabel;
+@property (weak, nonatomic) IBOutlet ABKLabel *inAppMessageHeaderLabel;
 @property (weak, nonatomic, nullable) IBOutlet UIImageView *graphicImageView;
 
 - (IBAction)dismissInAppMessage:(id)sender;