+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageViewController.h
@@ -23,9 +23,9 @@ NS_ASSUME_NONNULL_BEGIN
 @interface ABKInAppMessageViewController : UIViewController
 
 @property (strong) ABKInAppMessage *inAppMessage;
-@property (weak) IBOutlet UIImageView *iconImageView;
-@property (weak) IBOutlet UILabel *iconLabelView;
-@property (weak) IBOutlet ABKLabel *inAppMessageMessageLabel;
+@property (weak, nonatomic) IBOutlet UIImageView *iconImageView;
+@property (weak, nonatomic) IBOutlet UILabel *iconLabelView;
+@property (weak, nonatomic) IBOutlet ABKLabel *inAppMessageMessageLabel;
 
 /*
  * The initWithInAppMessage method may be used to pass the inAppMessage property to any custom view controller that you create.