+++ Example/Stopwatch/AppDelegate.h
@@ -1,13 +1,15 @@
 #import <UIKit/UIKit.h>
-#import "ABKPushURIDelegate.h"
 #import <UserNotifications/UserNotifications.h>
 #import "ABKInAppMessageControllerDelegate.h"
+#import "ABKURLDelegate.h"
 
 static NSString *const OverrideApiKeyStorageKey = @"com.appboy.stopwatch.apikey";
 static NSString *const OverrideEndpointStorageKey = @"com.appboy.stopwatch.endpoint";
 static NSString *const SetInAppMessageControllerDelegateKey = @"com.appboy.stopwatch.iamdelegate";
+static NSString *const SetURLDelegateKey = @"com.appboy.stopwatch.urldelegate";
 
-@interface AppDelegate : UIResponder <UIApplicationDelegate, ABKPushURIDelegate, UNUserNotificationCenterDelegate, ABKInAppMessageControllerDelegate>
+@interface AppDelegate : UIResponder <UIApplicationDelegate, UNUserNotificationCenterDelegate,
+ABKInAppMessageControllerDelegate, ABKURLDelegate>
 
 @property (strong, nonatomic) UIWindow *window;